feat : AnimalController에 작성한 기능 병합했습니다. #4
