+++ nyang/src/main/java/com/project/nyang/modules/animal/controller/AnimalController.java
@@ -1,5 +1,7 @@
 package com.project.nyang.modules.animal.controller;
 
+import com.project.nyang.global.common.api.ApiResponse;
+import com.project.nyang.global.common.api.ApiSuccessResponse;
 import com.project.nyang.modules.animal.dto.AnimalDTO;
 import com.project.nyang.modules.animal.dto.AnimalListDTO;
 import com.project.nyang.modules.animal.service.AnimalService;
@@ -10,6 +12,7 @@ import lombok.*;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDate;
@@ -57,5 +60,13 @@ public class AnimalController {
     ){
         PageRequest pageable = PageRequest.of(page, size);
         return animalService.getFilterAnimals(startDate,endDate,upKindCd,kindCd,regionCode,subRegionCode,pageable);
+    }
+
+
+    @Operation(summary = "상세 유기동물 조회", description = "유기동물 고유 번호를 바탕으로 해당 동물의 상세 정보를 조회합니다.")
+    @GetMapping("/{desertionNo}")
+    public ResponseEntity<ApiResponse<AnimalDTO>> getAnimalDetail(@PathVariable String desertionNo){
+        AnimalDTO animal = animalService.getAnimalDetail(desertionNo);
+        return ResponseEntity.ok(ApiSuccessResponse.success(animal,"동물정보 조회에 성공하였습니다"));
     }
 }
