+++ nyang/src/main/java/com/project/nyang/modules/animal/dto/AnimalDTO.java
@@ -1,9 +1,6 @@
 package com.project.nyang.modules.animal.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -12,7 +9,7 @@ import java.time.LocalDateTime;
  *
  * AnimalDTO 상세조회에 사용합니다.
  *
- * @author : 이지은
+ * @author : 이지은, 엄아영
  * @fileName : AnimalDTO
  * @since : 25. 7. 9.
  */
@@ -21,6 +18,7 @@ import java.time.LocalDateTime;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
+@Builder
 public class AnimalDTO {
     //유기 동물 번호
     private String desertionNo;
